requirements.txt
absl-py==2.3.1
addict==2.4.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
comm==0.2.3
configargparse==1.7.1
contourpy==1.3.0
cycler==0.12.1
dash==3.2.0
decorator==5.2.1
etils[epath,epy]==1.5.2
exceptiongroup==1.3.0
executing==2.2.1
farama-notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.19.1
flask==3.1.2
fonttools==4.60.1
fsspec==2025.9.0
glfw==2.10.0
gymnasium==0.29.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==8.7.0
importlib-resources==6.5.2
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kiwisolver==1.4.7
markupsafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.2.0
narwhals==2.7.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
open3d==0.19.0
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
plotly==6.3.1
proglog==0.1.12
prompt-toolkit==3.0.52
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyopengl==3.1.10
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.3
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
stable-baselines3==2.4.0
sb3-contrib==2.4.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.4.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.23.0
